requirements.txt
numpy~=1.26.4
opencv-python~=4.10.0.84
pillow~=10.4.0
pyzbar~=0.1.9
qreader~=3.14